Rename SymmetricKey.js -> symmetrickey.js
