+++ lib/backend/fipsrypto.js
@@ -8,7 +8,7 @@ const crypto = require('crypto');
 const fs = require('fs-extra');
 const sodium = require('sodium-native');
 const Util = require('../util');
-const SymmetricKey = require('./key/SymmetricKey');
+const SymmetricKey = require('./key/symmetrickey');
 const CryptoOperationException = require('../exception/cryptooperationexception');
 
 const MAGIC_HEADER = "fips:";
