+++ lib/backend/key/symmetrickey.js
