+++ lib/backend/moderncrypto.js
@@ -5,7 +5,7 @@ const base64url = require('rfc4648').base64url;
 const Backend =  require('../contract/backend');
 const sodium = require('sodium-native');
 const Util = require('../util');
-const SymmetricKey = require('./key/SymmetricKey');
+const SymmetricKey = require('./key/symmetrickey');
 const CryptoOperationException = require('../exception/cryptooperationexception');
 
 const MAGIC_HEADER = "nacl:";
