+++ lib/keyprovider/stringprovider.js
@@ -2,7 +2,7 @@
 
 const CryptoOperationException = require('../exception/cryptooperationexception');
 const KeyProvider = require('../contract/keyprovider');
-const SymmetricKey = require('../backend/key/SymmetricKey');
+const SymmetricKey = require('../backend/key/symmetrickey');
 const Util = require('../util');
 
 module.exports = class StringProvider extends KeyProvider
