+++ lib/util.js
@@ -3,7 +3,7 @@
 const crypto = require('crypto');
 const sodium = require('sodium-native');
 const arrayToBuffer = require('typedarray-to-buffer');
-const SymmetricKey = require('./backend/key/SymmetricKey');
+const SymmetricKey = require('./backend/key/symmetrickey');
 const CryptoOperationException = require('./exception/cryptooperationexception');
 
 /**
